Put each spec with the module it asserts
The fold left seven gameplay.spec.ts files testing things that are not
the rules — a pre-existing mix that renaming the files made visible.

Games whose bot lives in the game .tsx (ten-coins, number-covering,
discs-flip-or-remove, waning-stones, doubling-reduction, three-more) get
their strategy tests back in <game>.spec.ts, and cube-coloring's palette
test follows nodeColors into cube-coloring.spec.ts.

Three tests assert what the *bot* produces rather than what the rules
allow — bacteria's and take-and-point's "every move the bot offers is
legal", and modified-mill's symmetry canonicalisation, which is a search
optimisation and not a rule. Those move to bot-strategy.spec.ts.

Left alone: remove-row-or-column/multiple's generateStartBoard test reads
boardGrundy from the bot, but what it asserts is the start boards' ~50/50
balance, so it stays with the rules.

Test count is unchanged at 1306.
