+++ src/components/games/discs-flip-or-remove/discs-flip-or-remove.spec.ts
@@ -1,0 +1,60 @@
+import { describe, it, expect } from 'vitest';
+import { playBotMove } from '../../../test-utils';
+import { smartBotStrategy } from './discs-flip-or-remove';
+import { moves as gameMoves } from './gameplay';
+
+// board[0] = blue discs, board[1] = red discs.
+type Board = [number, number];
+
+const legalMoves = ([blue, red]: Board): Board[] => {
+  const next: Board[] = [];
+  if (blue >= 1) next.push([blue - 1, red]);
+  if (blue >= 2) next.push([blue - 2, red]);
+  if (red >= 1) next.push([blue + 1, red - 1]);
+  if (red >= 2) next.push([blue + 2, red - 2]);
+  return next;
+};
+
+const key = ([b, r]: Board) => `${b},${r}`;
+
+const botCandidates = (board: Board): Board[] => {
+  const seen = new Map<string, Board>();
+  for (let i = 0; i < 40; i++) {
+    const next = playBotMove(smartBotStrategy, gameMoves, board);
+    seen.set(key(next), next);
+  }
+  return [...seen.values()];
+};
+
+const moverWinsMemo = new Map<string, boolean>();
+const moverWins = (board: Board): boolean => {
+  const k = key(board);
+  const cached = moverWinsMemo.get(k);
+  if (cached !== undefined) return cached;
+  moverWinsMemo.set(k, false); // guard against cycles (there are none: total is non-increasing)
+  const wins = legalMoves(board).some(m => !moverWins(m));
+  moverWinsMemo.set(k, wins);
+  return wins;
+};
+
+describe('discs-flip-or-remove smart bot is optimal for both variants', () => {
+  for (const max of [6, 10]) {
+    it(`plays optimally across every position with up to ${max} discs`, () => {
+      for (let blue = 0; blue <= max; blue++) {
+        for (let red = 0; red + blue <= max; red++) {
+          const board: Board = [blue, red];
+
+          // The game value is exactly "blue is not a multiple of 3".
+          expect(moverWins(board), `game value at ${key(board)}`).toBe(blue % 3 !== 0);
+
+          if (!moverWins(board)) continue; // bot on the losing side: no win to protect
+          for (const next of botCandidates(board)) {
+            const isLegal = legalMoves(board).some(m => key(m) === key(next));
+            expect(isLegal, `${key(next)} legal from ${key(board)}`).toBe(true);
+            expect(moverWins(next), `bot blundered from ${key(board)} to ${key(next)}`).toBe(false);
+          }
+        }
+      }
+    });
+  }
+});
